+++ tests/3.0.x/__snapshots__/main/discriminatorOneOf/types/ElectricCar.ts
@@ -18,6 +18,8 @@ export type ElectricCar = (Vehicle & {
     type?: ElectricCarTypeEnumKey;
     /**
      * @description Battery capacity in kWh
+     *
+     * Format: `float`
      * @type number | undefined
     */
     batteryCapacity?: number;