+++ tests/3.0.x/__snapshots__/main/transformerStripDescriptions/types/GetItem.ts
@@ -6,6 +6,8 @@
 import type { Item } from "./Item.ts";
 
 /**
+ * @description
+ * Format: `int64`
  * @example test
  * @type integer
 */